+++ src/org/usfirst/frc/team6161/robot/commands/Intake.java
@@ -1,30 +1,39 @@
 package org.usfirst.frc.team6161.robot.commands;
 
+import org.usfirst.frc.team6161.robot.Robot;
 
+import edu.wpi.first.wpilibj.command.Command;
 
-import edu.wpi.first.wpilibj.command.CommandGroup;
+public class Intake extends Command{
 
+    public Intake() {
+    	requires(Robot.ArmsBase);
+        // Use requires() here to declare subsystem dependencies
+        // eg. requires(chassis);
+    }
 
-public class Intake extends CommandGroup {
-	 public Intake() {
-	        // Add Commands here:
-	        // e.g. addSequential(new Command1());
-	        //      addSequential(new Command2());
-	        // these will run in order.
+    // Called just before this Command runs the first time
+    protected void initialize() {
+    }
 
-	        // To run multiple commands at the same time,
-	        // use addParallel()
-	        // e.g. addParallel(new Command1());
-	        //      addSequential(new Command2());
-	        // Command1 and Command2 will run in parallel.
+    // Called repeatedly when this Command is scheduled to run
+    protected void execute() {
+    	Robot.ArmsBase.Intake();
+    }
 
-	        // A command group will require all of the subsystems that each member
-	        // would require.
-	        // e.g. if Command1 requires chassis, and Command2 requires arm,
-	        // a CommandGroup containing them would require both the chassis and the
-	        // arm.
-	    	addParallel(new RightArmIntake());
-	    	addParallel(new LeftArmIntake());
-	    
-	    }
+    // Make this return true when this Command no longer needs to run execute()
+    protected boolean isFinished() {
+        return false;
+    }
+
+    // Called once after isFinished returns true
+    protected void end() {
+    	Robot.ArmsBase.ArmStop();
+    }
+
+    // Called when another command which requires one or more of the same
+    // subsystems is scheduled to run
+    protected void interrupted() {
+    	end();
+    }
 }
