+++ src/org/usfirst/frc/team6161/robot/subsystems/SliderBase.java
@@ -2,9 +2,10 @@ package org.usfirst.frc.team6161.robot.subsystems;
 
 import org.usfirst.frc.team6161.robot.RobotMap;
 
-
+import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj.SpeedController;
 import edu.wpi.first.wpilibj.command.Subsystem;
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
 public class SliderBase extends Subsystem {
 
@@ -16,20 +17,37 @@ public class SliderBase extends Subsystem {
 	private SpeedController VerticalMotor;
 	private SpeedController HorzontalMotor;
 	
+	DigitalInput topVerticalLimitSwitch = new DigitalInput(1);
+	DigitalInput botVerticalLimitSwitch = new DigitalInput(2);
+	
+	DigitalInput frontHorizontalLimitSwitch = new DigitalInput(11);
+	DigitalInput rearHorizontalLimitSwitch = new DigitalInput(13);
+	
 	
 	public void init(){
 		VerticalMotor = RobotMap.sliderBaseVerticalMotor;
 		HorzontalMotor = RobotMap.sliderBaseHorizontalMotor;
+		
+		 
 	}
 	
     public void initDefaultCommand() {
         // Set the default command for a subsystem here.
         //setDefaultCommand(new MySpecialCommand());
     }
+    
+    public boolean isAtTop(){
+    	return !topVerticalLimitSwitch.get();
+ //   	return false;
+    }
     public void VerticalUp(){
-    	VerticalMotor.set(VerticalSpeed);
+   		VerticalMotor.set(VerticalSpeed);
+    	
+    	
+    	
     }
     public void VerticalDown(){
+    	SmartDashboard.putBoolean("Bottom HE sensor", botVerticalLimitSwitch.get());
     	VerticalMotor.set(-VerticalSpeed);
     }
     public void VerticalStop(){
